+++ src/controllers/RemoteBackupController.php
@@ -25,7 +25,7 @@ class RemoteBackupController extends Controller
 
     public function requirePluginConfigured()
     {
-        if (!RemoteBackup::getInstance()->getSettings()->configured()) {
+        if (!RemoteBackup::getInstance()->remotebackup->isConfigured()) {
             throw new BadRequestHttpException('Plugin is not correctly configured');
         }
     }